refactor: content-handlers, content-usecasesをclean-architecture構造に移動
srcのフォルダ構造に合わせて以下のように移動:
- content-handlers.js → clean-architecture/infrastructure/browser/handlers/content.js
- content-usecases.js → clean-architecture/application/usecases/contentOnMessageReceived.js

空になったrestricted-imports/ディレクトリを削除。
