+++ host-frontend-root/frontend-src-root/eslint-rules/clean-architecture/application/usecases/contentOnMessageReceived.js
